u-boot-tegra: Fix Nano not booting on 32.5.1 with custom dtb
u-boot was overriding bootargs set by pxe with the
ones in the fdt passed by cboot. This happened only
when a custom dtb was getting loaded, and the board
was failing to boot into the OS because the root was not
present in the kernel cmdline. We drop copying the
bootargs carveout and continue to use the args
as appended by extlinux.conf, which already contain
the cboot cmdline elements.

We drop this for the tx2 too since it was likely affected
too.

Signed-off-by: Alexandru Costache <alexandru@balena.io>