u-boot: Avoid kernel panic on L4T 32.5.1 TX2 rollback
As explaned in u-boot upstream commit f71f46f1dc
the DTB in RAM (CBoot DTB) has some new/added nodes
and properties that need to be added/updated to
any 'disk' kernel DTB (from extlinux.conf)
that gets loaded and handed to the kernel. However,
these 32.5.1 dtb changes cause a kernel panic if
loading an older 32.4.4 kernel, like in the case of
a rollback. Therefore, we avoid applying some of
the new or updated nodes in the case of rollback.

Signed-off-by: Alexandru Costache <alexandru@balena.io>