jetson-tx2-4-gb: Rename to jetson-tx2-4gb
Starting with L4T 32.5.1 the upstream yocto
machine for the TX2 4GB has been renamed to
jetson-tx2-devit-4gb. Using the previous 32.4.4 name
jetson-tx2-4gb would have forced us to rewrite all
the tx2 specfic code in our recipes, therefore
we created a new machine which ended in "4-gb"
so that we can re-use all the tx2 code and only
override what was specific to the 4GB version.

Now we can name this device type to jetson-tx2-4gb,
provided that we do not release this device type
in the "4-gb" version in the API.

NOTE: The boot0_bindiff for the TX2 4GB will now
need to be updated to match 32.5.1.

Signed-off-by: Alexandru Costache <alexandru@balena.io>