linux-tegra: Add temporary workaround for TX2 NX hang during boot
There appears to be a bug in u-boot which causes the kernel
to hang right after boot, if a dtb is loaded with u-boot.

Remove the FDT entry from extlinux.conf for the TX2 NX
so that the dtb is loaded solely from the raw partition,
until this is fixed in upstream tegra u-boot.

Signed-off-by: Alexandru Costache <alexandru@balena.io>